Refactor repo component
