+++ app/components/header.tsx
@@ -5,10 +5,10 @@ import styles from "~/styles/layout.module.css";
 export default function Header() {
   return (
     <div className={[styles.content, styles.column].join(" ")}>
-      <Title order={4} component="p">
+      <Title order={4} component="h1">
         Blake Netzeband
       </Title>
-      <Title order={4} component="p">
+      <Title order={4} component="h2">
         Full Stack Developer
       </Title>
     </div>
